refactor: avoid splitting vendor chunk by default
